pyproject.toml
[project]
name = "FinDL"
version = "0.0.1"
description = "Describe Your Cool Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torchvision>=0.15.0",
    "lightning>=2.0.0",
    "torchmetrics>=0.11.4",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "hydra-optuna-sweeper==1.2.0",
    "rootutils",
    "pre-commit",
    "rich",
    "pytest",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "akshare>=1.18.19",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "polars>=1.37.1",
    "xlsxwriter>=3.2.9",
    "pyarrow>=23.0.0",
    "tensorboard>=2.20.0",
    "fastexcel>=0.19.0",
    "scrapy>=2.11.0",
    "scrapy-fake-useragent>=1.4.4",
    "duckdb>=1.4.4",
    "jmespath>=1.1.0",
    "aiolimiter>=1.2.1",
    "curl-cffi>=0.14.0",
    "ruff>=0.14.14",
    "tenacity>=9.1.2",
    "pylance>=1.0.4",
    "neuralforecast>=3.1.4",
    "ray[tune]>=2.53.0",
    "pytorch-lightning>=2.6.0",
]

[project.scripts]
train_command = "src.train:main"
eval_command = "src.eval:main"

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
#url = "http://mirrors.cloud.aliyuncs.com/pypi/simple/" # 内网
default = true

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
